fix(blur): many small issues.
