+++ react/features/blur/actions.js
@@ -1,49 +1,48 @@
 // @flow
 
-import { getJitsiMeetGlobalNS } from '../base/util';
 import { getLocalVideoTrack } from '../../features/base/tracks';
 
-import {
-    BLUR_DISABLED,
-    BLUR_ENABLED
-} from './actionTypes';
+import { BLUR_DISABLED, BLUR_ENABLED } from './actionTypes';
+import { getBlurEffect } from './functions';
 
 const logger = require('jitsi-meet-logger').getLogger(__filename);
 
 /**
-* Signals the local participant is switching between blurred or
-* non blurred video.
+* Signals the local participant is switching between blurred or non blurred video.
 *
-* @param {boolean} enabled - If true enables video blur, false otherwise
-*
+* @param {boolean} enabled - If true enables video blur, false otherwise.
 * @returns {Promise}
 */
 export function toggleBlurEffect(enabled: boolean) {
     return function(dispatch: (Object) => Object, getState: () => any) {
-        if (getState()['features/blur'].blurEnabled !== enabled) {
-            const videoTrack = getLocalVideoTrack(getState()['features/base/tracks']).jitsiTrack;
+        const state = getState();
 
-            return getJitsiMeetGlobalNS().effects.createBlurEffect()
+        if (state['features/blur'].blurEnabled !== enabled) {
+            const { jitsiTrack } = getLocalVideoTrack(state['features/base/tracks']);
+
+            return getBlurEffect()
                 .then(blurEffectInstance =>
-                    videoTrack.enableEffect(enabled, blurEffectInstance)
+                    jitsiTrack.setEffect(enabled ? blurEffectInstance : undefined)
                         .then(() => {
                             enabled ? dispatch(blurEnabled()) : dispatch(blurDisabled());
                         })
                         .catch(error => {
                             enabled ? dispatch(blurDisabled()) : dispatch(blurEnabled());
-                            logger.log('enableEffect failed with error:', error);
+                            logger.error('setEffect failed with error:', error);
                         })
                 )
                 .catch(error => {
                     dispatch(blurDisabled());
-                    logger.log('createBlurEffect failed with error:', error);
+                    logger.error('getBlurEffect failed with error:', error);
                 });
         }
+
+        return Promise.resolve();
     };
 }
 
 /**
- * Signals the local participant that the blur has been enabled
+ * Signals the local participant that the blur has been enabled.
  *
  * @returns {{
  *      type: BLUR_ENABLED
@@ -56,7 +55,7 @@ export function blurEnabled() {
 }
 
 /**
- * Signals the local participant that the blur has been disabled
+ * Signals the local participant that the blur has been disabled.
  *
  * @returns {{
  *      type: BLUR_DISABLED
