+++ react/features/blur/index.js
@@ -1,4 +1,5 @@
 export * from './actions';
 export * from './components';
+export * from './functions';
 
 import './reducer';
